Create `response` module
