+++ src/handler.rs
@@ -1,8 +1,6 @@
-use ocpp::call_error::CallError;
-use ocpp::call_result::CallResult;
-use std::{any::TypeId, marker::PhantomData};
-
 use crate::request::{FromRequest, Request};
+use crate::response::*;
+use std::{any::TypeId, marker::PhantomData};
 
 pub struct ConcreteHandler<F, Args, O> {
     func: F,
@@ -18,29 +16,6 @@ pub trait Handler {
     type Response;
     fn call(&self, args: &Request) -> Self::Response;
     fn routing_key(&self) -> TypeId;
-}
-
-pub trait IntoResponse {
-    fn into_response(&self) -> Result<CallResult, CallError>;
-}
-
-impl IntoResponse for CallResult {
-    fn into_response(&self) -> Result<CallResult, CallError> {
-        Ok(self.clone())
-    }
-}
-
-use ocpp::call_result::Payload;
-use ocpp::v16::authorize_response::AuthorizeResponse;
-
-impl IntoResponse for AuthorizeResponse {
-    fn into_response(&self) -> Result<CallResult, CallError> {
-        Ok(CallResult {
-            message_type_id: 3,
-            unique_id: "123".to_string(),
-            payload: Payload::Authorize(self.clone()),
-        })
-    }
 }
 
 impl<Func, A: FromRequest + 'static, O> Handler for ConcreteHandler<Func, (A,), O>
